backend/requirements.txt
fastapi==0.115.0
uvicorn==0.21.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
pydantic==2.9.2
psycopg[binary]==3.2.2